chore: rubocop fix and new module helper
