+++ app/helpers/application_helper.rb
@@ -126,16 +126,4 @@ module ApplicationHelper
   def staging_site?
     request.host.include?('staging')
   end
-
-  def middle_from_array(arr)
-    if arr.length.even?
-      arr[(arr.length / 2) - 1] + arr[arr.length / 2]
-    else
-      arr[arr.length / 2]
-    end
-  end
-
-  def middle_from_rel(rel)
-    rel.offset(rel.size / 2).first
-  end
 end
