+++ app/helpers/collection_helper.rb
@@ -1,0 +1,13 @@
+module CollectionHelper
+  def middle_from_array(arr)
+    if arr.length.even?
+      arr[(arr.length / 2) - 1] + arr[arr.length / 2]
+    else
+      arr[arr.length / 2]
+    end
+  end
+
+  def middle_from_rel(rel)
+    rel.offset(rel.size / 2).first
+  end
+end
